lib/modules: disallowedOptions, fix _file
Adds an argument to evalModules for disallowedOptions. Useful to
prevent declaring of options named `config` and `options`. Should
probably be `extraAllowedOptions` later, with `config` and `options`
(and some more such as `imports`, etc.) the
only one being disallowed.

Fixes the file/_file thing, which prevented unifyModuleSyntax from
returning a reusable module. Before unifyModuleSyntax took a module of
the form { _file = ...; } and returned { file = ...; }, but not
accepting such an argument again.