nixos/system: Better type system.build
The previous types.attrs type doesn't allow overriding or other features of the
module system. Using this new type means e.g. the following can be used
to change system.build.binsh reliably (let's ignore the
environment.binsh option for now):

  system.build.binsh = mkForce pkgs.bashInteractive_5;

While you could assign

  system.build.binsh = pkgs.bashInteractive_5;

previously, assigning this multiple times would make the actual
definition to be used dependent on the imports ordering.

As such, this change is also not fully backwards compatible, since
assigning a single option with the same priority multiple times now
leads to an error, but that's how it's meant to be.