nixos/zsh: Set aliases before running interactiveShellInit
Aliases are just a text -> text mapping, and it therefore doesn't matter
when they are defined. The interactive shell init however may depend on
variables existing at the time of it running. And if aliases were only
defined after the shell init, it could not refer to any aliases.

E.g. previously if you had

  shellAliases.foo = "echo foo";
  interactiveShellInit = ''
    fooFun() {
      foo
    }
  ''

This would first run the shell init, which resolves `foo` to be an
unknown command. Running `fooFun` therefore produces an error, even if
`foo` is defined in an alias later

In comparison, after this change the alias is defined before the shell
init, making it resolve correctly, allowing `fooFun` to output `foo`
without error.